updated docs, pre/post/setup/teardown, direct exec jobs
jobs can be directly executed in a shell with 'job.exec'.

pre and post can be ignored on a job-by-job basis with 'job.nopre'
and 'job.nopost'.

setup and teardown can now either be direct exec w/ 'job.exec' or
scheduled. if scheduled, setup is a dep for all other jobs and teardown
depends on all other jobs.